net/netutil: move some net utils from control/controlhttp to netutil
In prep for reuse elsewhere.
